Add WIP Algorand `TokenBridge` protocol
Work in progress implementation of the `TokenBridge` protocol
for the Algorand blockchain. Based on functionality from:
- https://github.com/wormhole-foundation/wormhole/tree/38f02fa2fc8e80fc3af213959fe397cc6a936778/sdk/js/src/algorand
- https://github.com/wormhole-foundation/wormhole/tree/38f02fa2fc8e80fc3af213959fe397cc6a936778/sdk/js/src/token_bridge

The current implementation has not been tested. The
`getWrappedAsset()` has no implementation yet.

Current implementation changes the API of the
`core/definitions/src/protocols/tokenBridge.ts` which might
need to be reverted/changed in a future commit.