Fix wrapped token redemption on Algorand
Fix the implementation of `*redeem()` for wrapped tokens on
Algorand.

Current implementation uses `serialize(vaa)` to create a Uint8Array
with the content of the VAA, that can be then parsed by
`_parseVAAAlgorand()` and create the same `ParsedVAA` that
the `wormhole-sdk` uses [0].

The code should/would be updated to use the
`vaa: TokenBridge.VAA<'Transfer' | 'TransferWithPayload'>` type
passed as an argument to `*redeem()`.

[0]: https://github.com/wormhole-foundation/wormhole/blob/38f02fa2fc8e80fc3af213959fe397cc6a936778/sdk/js/src/algorand/Algorand.ts#L244C8-L244C22