isSupported only checks chains & tokens
This moves `amount` out of RouteTransferRequest, and makes it
a parameter passed into quote(), validate(), and initiate().

The idea is that a bridging interface should show you all available
routes given only the chains and token(s) you are dealing with.
Once you see the list, you can tweak the amount, and any other
options like native gas dropoff and slippage. Based on your amount
certain supported routes may not be available because the amount is
either too large or too small, but it's useful to see this as a
user because you may decide to adjust your amount in order to get access
to a particular route.

This change will also mean we don't have to call `resolve` every time
the user tweaks their amount. We only need to call it if the chains or
tokens change.