AutomaticTokenBridge gas dropoff percentage of max change and decimal scaling fix
Previously the AutomaticTokenBridge route calculated the gas dropoff amount
as a percentage of the input amount. However this has been changed to
be a percentage of the maxSwapAmount as it's an easier to use interface
for users of the library (e.g. Connect).

Also fixed an issue where the native gas wasn't scaled correctly
when the source and destination chains have different decimals
for the input token.